Move bul-storage off into its own file
