+++ opencog/persist/sql/multi-driver/SQLAtomStorage.cc
@@ -242,8 +242,6 @@ void SQLAtomStorage::extract_callback(const AtomPtr& atom)
 
 /* ================================================================== */
 /* AtomTable UUID stuff */
-#define BUFSZ 250
-
 void SQLAtomStorage::store_atomtable_id(const AtomTable& at)
 {
 	UUID tab_id = at.get_uuid();
@@ -260,6 +258,7 @@ void SQLAtomStorage::store_atomtable_id(const AtomTable& at)
 		store_atomtable_id(*env);
 	}
 
+#define BUFSZ 80
 	char buff[BUFSZ];
 	snprintf(buff, BUFSZ,
 		"INSERT INTO Spaces (space, parent) VALUES (%ld, %ld);",
@@ -269,51 +268,7 @@ void SQLAtomStorage::store_atomtable_id(const AtomTable& at)
 	rp.exec(buff);
 }
 
-
-/* ================================================================ */
-
-#define STMT(colname,val) { \
-	if (notfirst) { cols += ", "; vals += ", "; } else notfirst = true; \
-	cols += colname; \
-	vals += val; \
-}
-
-#define STMTI(colname,ival) { \
-	char buff[BUFSZ]; \
-	snprintf(buff, BUFSZ, "%d", ival); \
-	STMT(colname, buff); \
-}
-
-#define STMTF(colname,fval) { \
-	char buff[BUFSZ]; \
-	snprintf(buff, BUFSZ, "%22.16g", fval); \
-	STMT(colname, buff); \
-}
-
 /* ================================================================== */
-/**
- * Return largest distance from this atom to any node under it.
- * Nodes have a height of 0, by definition.  Links that contain only
- * nodes in their outgoing set have a height of 1, by definition.
- * The height of a link is, by definition, one more than the height
- * of the tallest atom in its outgoing set.
- * @note This can conversely be viewed as the depth of a tree.
- */
-int SQLAtomStorage::get_height(const Handle& atom)
-{
-	if (not atom->is_link()) return 0;
-
-	int maxd = 0;
-	for (const Handle& h : atom->getOutgoingSet())
-	{
-		int d = get_height(h);
-		if (maxd < d) maxd = d;
-	}
-	return maxd + 1;
-}
-
-/* ================================================================ */
-
 /// Return the UUID of the handle, if it is known.
 /// If the handle is in the database, then the correct UUID is returned.
 /// If the handle is NOT in the database, this returns the invalid UUID.
@@ -380,253 +335,6 @@ UUID SQLAtomStorage::get_uuid(const Handle& h)
 void SQLAtomStorage::flushStoreQueue()
 {
 	_write_queue.barrier();
-}
-
-/* ================================================================ */
-/**
- * Retreive the incoming set of the indicated atom.
- */
-void SQLAtomStorage::getIncoming(AtomTable& table, const char *buff)
-{
-	std::vector<PseudoPtr> pset;
-	Response rp(conn_pool);
-	rp.store = this;
-	rp.height = -1;
-	rp.pvec = &pset;
-	rp.exec(buff);
-	rp.rs->foreach_row(&Response::fetch_incoming_set_cb, &rp);
-
-	HandleSeq iset;
-	std::mutex iset_mutex;
-
-	// Parallelize always.
-	opencog::setting_omp(NUM_OMP_THREADS, NUM_OMP_THREADS);
-
-	// A parallel fetch is much much faster, esp for big osets.
-	// std::for_each(std::execution::par_unseq, ... requires C++17
-	OMP_ALGO::for_each(pset.begin(), pset.end(),
-		[&] (const PseudoPtr& p)
-	{
-		Handle hi(get_recursive_if_not_exists(p));
-		hi = table.add(hi, false);
-		_tlbuf.addAtom(hi, p->uuid);
-		get_atom_values(hi);
-		std::lock_guard<std::mutex> lck(iset_mutex);
-		iset.emplace_back(hi);
-	});
-
-#ifdef STORAGE_DEBUG
-	_num_get_insets++;
-	_num_get_inlinks += iset.size();
-#endif // STORAGE_DEBUG
-}
-
-/**
- * Retreive the entire incoming set of the indicated atom.
- */
-void SQLAtomStorage::getIncomingSet(AtomTable& table, const Handle& h)
-{
-	// If the uuid is not known, then the atom is not in storage,
-	// and therefore, cannot have an incoming set.  Just return.
-	UUID uuid = check_uuid(h);
-	if (TLB::INVALID_UUID == uuid) return;
-
-	char buff[BUFSZ];
-	snprintf(buff, BUFSZ,
-		"SELECT * FROM Atoms WHERE outgoing @> ARRAY[CAST(%lu AS BIGINT)];",
-		uuid);
-
-	// Note: "select * from atoms where outgoing@>array[556];" will
-	// return all links with atom 556 in the outgoing set -- i.e. the
-	// incoming set of 556.  We could also use && here instead of @>
-	// but I don't know if this one is faster.
-	// The cast to BIGINT is needed, as otherwise one gets
-	// ERROR:  operator does not exist: bigint[] @> integer[]
-
-	getIncoming(table, buff);
-}
-
-/**
- * Retreive the incoming set of the indicated atom, but only those atoms
- * of type t.
- */
-void SQLAtomStorage::getIncomingByType(AtomTable& table, const Handle& h, Type t)
-{
-	// If the uuid is not known, then the atom is not in storage,
-	// and therefore, cannot have an incoming set.  Just return.
-	UUID uuid = check_uuid(h);
-	if (TLB::INVALID_UUID == uuid) return;
-
-	int dbtype = storing_typemap[t];
-
-	char buff[BUFSZ];
-	snprintf(buff, BUFSZ,
-		"SELECT * FROM Atoms WHERE type = %d AND outgoing @> ARRAY[CAST(%lu AS BIGINT)];",
-		dbtype, uuid);
-
-	getIncoming(table, buff);
-}
-
-/* ================================================================ */
-
-void SQLAtomStorage::load(AtomTable &table)
-{
-	UUID max_nrec = reserve();
-	_load_count = 0;
-	max_height = getMaxObservedHeight();
-	printf("Loading all atoms; maxuuid=%lu max height=%d\n",
-		max_nrec, max_height);
-	bulk_load = true;
-	bulk_start = time(0);
-
-	setup_typemap();
-
-#define NCHUNKS 300
-#define MINSTEP 10123
-	std::vector<unsigned long> steps;
-	unsigned long stepsize = MINSTEP + max_nrec/NCHUNKS;
-	for (unsigned long rec = 0; rec <= max_nrec; rec += stepsize)
-		steps.push_back(rec);
-
-	printf("Loading all atoms: "
-		"Max Height is %d stepsize=%lu chunks=%zu\n",
-		 max_height, stepsize, steps.size());
-
-	// Parallelize always.
-	opencog::setting_omp(NUM_OMP_THREADS, NUM_OMP_THREADS);
-
-	for (int hei=0; hei<=max_height; hei++)
-	{
-		unsigned long cur = _load_count;
-
-		OMP_ALGO::for_each(steps.begin(), steps.end(),
-			[&](unsigned long rec)
-		{
-			Response rp(conn_pool);
-			rp.table = &table;
-			rp.store = this;
-			char buff[BUFSZ];
-			snprintf(buff, BUFSZ, "SELECT * FROM Atoms WHERE "
-			         "height = %d AND uuid > %lu AND uuid <= %lu;",
-			         hei, rec, rec+stepsize);
-			rp.height = hei;
-			rp.exec(buff);
-			rp.rs->foreach_row(&Response::load_all_atoms_cb, &rp);
-		});
-		printf("Loaded %lu atoms at height %d\n", _load_count - cur, hei);
-	}
-
-	time_t secs = time(0) - bulk_start;
-	double rate = ((double) _load_count) / secs;
-	printf("Finished loading %lu atoms in total in %d seconds (%d per second)\n",
-		(unsigned long) _load_count, (int) secs, (int) rate);
-	bulk_load = false;
-
-	// synchrnonize!
-	table.barrier();
-}
-
-void SQLAtomStorage::loadType(AtomTable &table, Type atom_type)
-{
-	UUID max_nrec = reserve();
-	_load_count = 0;
-
-	// For links, assume a worst-case height.
-	// For nodes, its easy ... max_height is zero.
-	if (nameserver().isNode(atom_type))
-		max_height = 0;
-	else
-		max_height = getMaxObservedHeight();
-
-	setup_typemap();
-	int db_atom_type = storing_typemap[atom_type];
-
-#define NCHUNKS 300
-#define MINSTEP 10123
-	std::vector<unsigned long> steps;
-	unsigned long stepsize = MINSTEP + max_nrec/NCHUNKS;
-	for (unsigned long rec = 0; rec <= max_nrec; rec += stepsize)
-		steps.push_back(rec);
-
-	logger().debug("SQLAtomStorage::loadType: "
-		"Max Height is %d stepsize=%lu chunks=%lu\n",
-		 max_height, stepsize, steps.size());
-
-	// Parallelize always.
-	opencog::setting_omp(NUM_OMP_THREADS, NUM_OMP_THREADS);
-
-	for (int hei=0; hei<=max_height; hei++)
-	{
-		unsigned long cur = _load_count;
-
-		OMP_ALGO::for_each(steps.begin(), steps.end(),
-			[&](unsigned long rec)
-		{
-			Response rp(conn_pool);
-			rp.table = &table;
-			rp.store = this;
-			char buff[BUFSZ];
-			snprintf(buff, BUFSZ, "SELECT * FROM Atoms WHERE type = %d "
-			         "AND height = %d AND uuid > %lu AND uuid <= %lu;",
-			         db_atom_type, hei, rec, rec+stepsize);
-			rp.height = hei;
-			rp.exec(buff);
-			rp.rs->foreach_row(&Response::load_if_not_exists_cb, &rp);
-		});
-		logger().debug("SQLAtomStorage::loadType: "
-		               "Loaded %lu atoms of type %d at height %d\n",
-			_load_count - cur, db_atom_type, hei);
-	}
-	logger().debug("SQLAtomStorage::loadType: Finished loading %lu atoms in total\n",
-		(unsigned long) _load_count);
-
-	// Synchronize!
-	table.barrier();
-}
-
-/// Store all of the atoms in the atom table.
-void SQLAtomStorage::store(const AtomTable &table)
-{
-	max_height = 0;
-	_store_count = 0;
-
-#ifdef ALTER
-	rename_tables();
-	create_tables();
-#endif
-
-	UUID max_uuid = _tlbuf.getMaxUUID();
-	printf("Max UUID is %lu\n", max_uuid);
-
-	// If we are storing to an absolutely empty database, then
-	// skip all UUID lookups completely!  This is not a safe
-	// operation for non-empty databases, but has a big performance
-	// impact for clean stores.
-	// uuid==1 is PredicateNode TruthValueKey
-	// uuid==2 is unissued.
-	if (2 >= max_uuid) bulk_store = true;
-
-	setup_typemap();
-	store_atomtable_id(table);
-
-	bulk_start = time(0);
-
-	// Try to knock out the nodes first, then the links.
-	table.foreachHandleByType(
-		[&](const Handle& h)->void { storeAtom(h); },
-		NODE, true);
-
-	table.foreachHandleByType(
-		[&](const Handle& h)->void { storeAtom(h); },
-		LINK, true);
-
-	flushStoreQueue();
-	bulk_store = false;
-
-	time_t secs = time(0) - bulk_start;
-	double rate = ((double) _store_count) / secs;
-	printf("\tFinished storing %lu atoms total, in %d seconds (%d per second)\n",
-		(unsigned long) _store_count, (int) secs, (int) rate);
 }
 
 /* ================================================================ */
